+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Return correct webId and podUrl for subdomain pods from `v7.createAccount`.
+
 ## [0.1.0] - 2025-01-30
 
 ### Added